Make submit button fixed positioned under certain conditions; improve usability of tabs.
Conditions are: the endpoint’s form exceeds the viewport, the viewport
is taller than 545px, the viewport is wider than 1000px, and the browser
supports CSS calc (and has JS enabled so that this can be detected). If
any of these conditions are not met, the button will remain in its
original position.