Remove autofocus on endpoint form.
Autofocus should only be used when there is near absolute certainty that
the user will take a certain action, and even then only sparingly. In
this case, the user may not necessarily wish to use the form, she could
merely be seeking information from it. Even if she does use the form
she may not always wish to give a value to first field. It’s better not
to second guess the user in this case.